refactor(ui): move components to correct tiers
