+++ src/app/(main)/(my-order)/my-orders/[orderId]/invitation/_components/CoupleInfoSection.tsx
@@ -2,8 +2,7 @@
 
 import { Card, CardContent, CardHeader, CardTitle, TypographyH3 } from "@/ui/components/atoms";
 
-import { BankField } from "@/ui/components/molecules";
-import { TextField } from "@/ui/components/organisms";
+import { BankField, TextField } from "@/ui/components/organisms";
 
 import type { InvitationContent } from "@/core/domain";
 import type { BanksResponse } from "@/core/schemas";
