+++ src/app/(main)/(my-order)/my-orders/[orderId]/invitation/_components/ParentsInfoSection.tsx
@@ -2,8 +2,7 @@
 
 import { Card, CardContent, CardHeader, CardTitle, Collapsible, CollapsibleContent, CollapsibleTrigger, TypographyH3, TypographySmall } from "@/ui/components/atoms";
 
-import { BankField } from "@/ui/components/molecules";
-import { TextField } from "@/ui/components/organisms";
+import { BankField, TextField } from "@/ui/components/organisms";
 
 import { ChevronDown } from "lucide-react";
