+++ src/ui/components/molecules/index.ts
@@ -1,9 +1,7 @@
 export * from "./Alert";
 export * from "./AutoCompleteList";
-export * from "./BankField";
 export * from "./BaseSelect";
 export * from "./CloudImage";
 export * from "./CursorPagination";
-export * from "./LegalDocumentTemplate";
 export * from "./Pricing";
 export * from "./ProductCard";
