+++ src/ui/components/templates/LegalDocumentTemplate.tsx
