+++ src/ui/components/templates/index.ts
@@ -1,0 +1,1 @@
+export * from "./LegalDocumentTemplate";
